Moved tests outside of src
