+++ tests/sensor_definition.json
