+++ tests/test_amplifier.py
@@ -1,17 +1,19 @@
 import unittest
 from its_preselector.web_relay_preselector import WebRelayPreselector
 import json
+from pathlib import Path
 
 
 class TestAmplifier(unittest.TestCase):
 
     @classmethod
     def setUpClass(cls):
-        file = open('test_metadata.sigmf-meta')
+        fpath = Path(__file__).parent.resolve()
+        file = open(fpath / 'test_metadata.sigmf-meta')
         sensor_def = json.load(file)
         file.close()
         cls.preselector = WebRelayPreselector(sensor_def, {})
-        null_file = open('null_preselector.sigmf-meta')
+        null_file = open(fpath / 'null_preselector.sigmf-meta')
         null_def = json.load(null_file)
         null_file.close()
         cls.empty_preselector = WebRelayPreselector(null_def, {})
